refactor: remove delete dropzone
