+++ project-canvas/src/components/StoryMapView/Components/DeleteButton.tsx
